refactoring to put joi validator into utilities; removing entity that is not really an entity
